# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SECOM MES Database API",
    "description": "REST API for SECOM Manufacturing Execution System Database\n\nThis API provides access to semiconductor manufacturing data including:\n- Master data (Equipment, Shifts, Operators, Product Types, Features)\n- Production data (Lots, Measurements, Quality Results)\n- Analytics (Production summaries, Equipment health, Shift performance, Quality analytics)\n\n**Database**: MariaDB 11.2\n**Total Records**: 1,567 lots with 590 sensor measurements each\n**Time Period**: July 2008 - September 2008\n\n**Key Endpoints**:\n- `/equipment` - Manufacturing equipment management\n- `/lots` - Production lot tracking with filters\n- `/measurements` - Sensor measurement data\n- `/quality` - Quality inspection results\n- `/analytics` - Analytics and reporting endpoints\n",
    "contact": {
      "name": "SECOM MES Team",
      "email": "mes@secom.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/equipment/{equipmentId}": {
      "get": {
        "tags": [
          "Equipment"
        ],
        "summary": "Get equipment by ID",
        "description": "Returns a single equipment by equipment ID",
        "operationId": "getEquipmentById",
        "parameters": [
          {
            "name": "equipmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Equipment"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Equipment"
        ],
        "summary": "Update equipment",
        "description": "Updates an existing equipment record",
        "operationId": "updateEquipment",
        "parameters": [
          {
            "name": "equipmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Equipment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Equipment"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Equipment"
        ],
        "summary": "Delete equipment",
        "description": "Deletes an equipment record",
        "operationId": "deleteEquipment",
        "parameters": [
          {
            "name": "equipmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/equipment": {
      "get": {
        "tags": [
          "Equipment"
        ],
        "summary": "Get all equipment",
        "description": "Returns list of all manufacturing equipment",
        "operationId": "getAllEquipment",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Equipment"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Equipment"
        ],
        "summary": "Create new equipment",
        "description": "Creates a new equipment record",
        "operationId": "createEquipment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Equipment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Equipment"
                }
              }
            }
          }
        }
      }
    },
...
    "/equipment/type/{type}": {
      "get": {
        "tags": [
          "Equipment"
        ],
        "summary": "Get equipment by type",
        "description": "Returns all equipment of a specific type (CVD, Etcher, etc.)",
        "operationId": "getEquipmentByType",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Equipment"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/equipment/code/{code}": {
      "get": {
        "tags": [
          "Equipment"
        ],
        "summary": "Get equipment by code",
        "description": "Returns equipment by unique equipment code",
        "operationId": "getEquipmentByCode",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Equipment"
                }
              }
            }
          }
        }
      }
    },
...
    "/analytics/high-risk-lots": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get high-risk lots above threshold",
        "description": "Returns lots with predicted risk above specified threshold with detailed information",
        "operationId": "getHighRiskLots",
        "parameters": [
          {
            "name": "threshold",
            "in": "query",
            "description": "Risk threshold (0.0 to 1.0)",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0.7
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit results (default 50)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HighRiskLotDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/analytics/feature-importance": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get feature importance rankings",
        "description": "Returns top features by importance score for a specific defect type or overall",
        "operationId": "getFeatureImportance",
        "parameters": [
          {
            "name": "defectType",
            "in": "query",
            "description": "Defect type (overall, electrical_fail, dimensional_oor, etc.)",
            "required": false,
            "schema": {
              "type": "string",
              "default": "overall"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit results (default 10)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Equipment": {
        "type": "object",
        "properties": {
          "equipmentId": {
            "type": "integer",
            "format": "int32"
          },
          "equipmentCode": {
            "type": "string"
          },
          "equipmentName": {
            "type": "string"
          },
          "equipmentType": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "manufacturer": {
            "type": "string"
          },
          "installDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "HighRiskLotDto": {
        "type": "object",
        "properties": {
          "lotNumber": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "equipmentCode": {
            "type": "string"
          },
          "predictedRisk": {
            "type": "number"
          },
          "riskScore": {
            "type": "number"
          },
          "testDatetime": {
            "type": "string",
            "format": "date-time"
          },
          "actualClassification": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}